pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
google-generativeai = "0.3.1"
replit-ai = "^0.0.11"
ipython = "^8.18.1"
xds-protos = "^1.60.0"
streamlit = "1.29.0"
langchain = "0.0.350"
streamlit-chat = "0.1.1"
st-multimodal-chatinput = "0.1.6"
langchain-core = "0.1.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"